Reduce test score section repetitions
The test score was very long, to test whether the rsound-composer scheduler could handle particularly long scores.  It handled them fine.  I am shortening the number of repetitions of  ```sect``` to make the file more legible, since the readme  will point to this file as an example until the documentation is finished.